Fixed DbSkseEvents
DbSkseEvents are no longer serialized as it was causing CTDs if fast travelling too fast on Skyrim AE. This means that mod authors will have to re-register the events on game load.
The OnLoadGameGlobal event now no longer has to be registered for, it is always sent so you can use it to re register the other events.
The OnHitGlobal event now sends the correct ammo if shooting your last arrow or bolt. Ammo is now tracked with the "bowDraw" animation event for actors.